# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Discutex API",
        "description": "API for managing threads, responses, and votes",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/threads": {
            "get": {
                "tags": [
                    "Threads"
                ],
                "summary": "Get all threads",
                "description": "Retrieves all threads in the system.",
                "operationId": "getAllThreads",
                "responses": {
                    "200": {
                        "description": "Threads found successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "threads": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Threads"
                ],
                "summary": "Create a new thread",
                "description": "Creates a new thread for the authenticated user.",
                "operationId": "createThread",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the thread",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Content of the thread",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thread created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "thread": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "content": {
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated."
                    },
                    "400": {
                        "description": "Bad request (missing title or content)."
                    }
                }
            }
        },
        "/threads/my": {
            "get": {
                "tags": [
                    "Threads"
                ],
                "summary": "Get threads of the authenticated user",
                "description": "Retrieves all threads created by the authenticated user.",
                "operationId": "getMyThreads",
                "responses": {
                    "200": {
                        "description": "Threads found successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "threads": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated."
                    }
                }
            }
        },
        "/threads/{id}": {
            "delete": {
                "tags": [
                    "Threads"
                ],
                "summary": "Delete a specific thread",
                "description": "Deletes a thread created by the authenticated user.",
                "operationId": "deleteThread",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the thread to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thread deleted successfully."
                    },
                    "404": {
                        "description": "Thread not found or not authorized to delete this thread."
                    },
                    "401": {
                        "description": "User not authenticated."
                    }
                }
            }
        },
...
    },
...
}